+++ src/js/admin/acf-field-group.js
@@ -4,28 +4,21 @@
 		selector = '[data-key="repeater_field"] select';
 
 	// reduce value & label field choices when repeater field changes
-	$(document).on( 'change', selector, function(e){
+	function setupRepeaterChoices( ) {
 		if ( !! repeated_fields[ $(this).val() ] ) {
 			var html = '',
 				repeater = $(this).val();
-			/*
-			$(this).closest('.acf-field-settings')
-				.find('[data-key="repeater_label_field"] select,[data-key="repeater_value_field"] select')
-				.each(function(i,el){
-					$(this).find('optgroup').each(function(i,el){
-						if ( $(this).is('[label="'+repeater+'"]') ) {
 
-						}
-
-					})
-				});
-			/*/
+			// generate options
 			$.each(repeated_fields[ repeater ],function( val, label ){
 				html += '<option value="'+val+'">' + label + '</option>';
 			});
+
+			// setup options
 			$(this).closest('.acf-field-settings')
 				.find('[data-key="repeater_label_field"] select,[data-key="repeater_value_field"] select')
 				.each(function(i,el){
+
 					var val = $(this).val(),
 						choiceNull = $(this).find('option[value=""]'),
 						field = acf.getField($(this).closest('.acf-field'));
@@ -37,10 +30,11 @@
 					$(this).val( val );
 
 				});
-			//*/
-
 		}
-	}).ready(function(){
-		$(selector).trigger('change');
-	});
+	}
+	$(document)
+		.on( 'change', selector, setupRepeaterChoices )
+		.ready(function(){
+			$(selector).each( setupRepeaterChoices );
+		});
 })(jQuery);