+++ CHANGELOG.md
@@ -4,6 +4,9 @@ ACF-WP-Objects Changelog
 0.1.14
 ------
  - Fix uninstall fatal
+ - Repeater-Choice: improve color select style
+ - Introduce Gulp
+ - Fix Issue in Field Group admin: acf noted changes but nothing changed
 
 0.1.13
 ------